refactor(trajectory_visualization): removed the moved trajectory visualization scripts from their old location
